Refactored client api files - Move auth routes and quiz routes to separate files - Moved BASE_URL to vite env
